feat: add service worker and port listeners, and implement fetch to convert url
Added a service worker script to allow fetching of shortened urls to be done with less restriction. Added port listeners on service worker and content scripts to communicate. Added a method in the service worker to convert shortened URLs to full URLs. Updated host permissions on manifest.json to all URLs to allow CORS (URL will be narrowed in the future)